Split tests after extracting class Game
